Allow only logged in users

+++ src/infrastructure/django_framework/camera_ctrl/views/app_logging/logs.py
@@ -3,12 +3,13 @@ from django.db.models.functions import Concat
 from django.shortcuts import redirect
 from django.urls import reverse_lazy
 from django.views.generic import ListView, DeleteView, View
+from django.contrib.auth.mixins import LoginRequiredMixin, AccessMixin
 
 from camera_ctrl.filters.app_logging import LogFilter
 from camera_ctrl.models import HistoryUnit
 
 
-class LogsList(ListView):
+class LogsList(LoginRequiredMixin, ListView):
     model = HistoryUnit
     template_name = 'app_logging/logs_list.html'
     paginate_by = 30
@@ -23,7 +24,7 @@ class LogsList(ListView):
         return filtered_logs.qs
 
 
-class LogsDelete(DeleteView):
+class LogsDelete(LoginRequiredMixin, DeleteView):
     model = HistoryUnit
     success_url = reverse_lazy('all_logs')
 
@@ -31,7 +32,10 @@ class LogsDelete(DeleteView):
         return self.delete(request, *args, **kwargs)
 
 
-class LogsDeleteAll(View):
+class LogsDeleteAll(AccessMixin, View):
     def dispatch(self, request, *args, **kwargs):
+        if not request.user.is_authenticated:
+            return self.handle_no_permission()
+
         HistoryUnit.objects.all().delete()
         return redirect(reverse_lazy('all_logs'))
